requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.1.0
Pillow==8.4.0
psycopg2==2.9.3
PyJWT==2.3.0
pymongo==4.0.1
pytz==2021.3
sqlparse==0.2.4
whitenoise==5.3.0
winkerberos==0.8.0
